feat: add view paths functionality to TUI for displaying config and log directory
