+++ .github/workflows/dotnet.yml
@@ -60,29 +60,3 @@ jobs:
 
     - name: Check for vulnerable packages
       run: dotnet list EasySave.sln package --vulnerable --include-transitive
-
-  publish:
-    name: Publish
-    runs-on: ubuntu-latest
-    needs: [ build-and-test, security ]
-    if: github.event_name == 'push' && github.ref == 'refs/heads/develop'
-
-    steps:
-    - uses: actions/checkout@v4
-
-    - name: Setup .NET
-      uses: actions/setup-dotnet@v4
-      with:
-        dotnet-version: 10.0.x
-
-    - name: Restore dependencies
-      run: dotnet restore EasySave.sln
-
-    - name: Publish application
-      run: dotnet publish EasySave.csproj -c Release -o ./publish
-
-    - name: Upload publish artifact
-      uses: actions/upload-artifact@v4
-      with:
-        name: EasySave-publish
-        path: publish/
